Wrote CheckForExistingSubmission() and debugged
The system will now check if there is an existing submission for the current date if the user tries to answer daily questions when there is already a entry in the database with that date. After the user answers the first question, the system will ask the user if they wish to overwrite the previous submission. If they answer yes, they can continue answering the questions and their submission will overwrite the previous one. If they select no, the window will close and the previous submission will remain untouched.