Added functions for deleting journal entries
in order for the app to display multiline journal entries, I had to change the listview to a textbox- only textboxes and richtextboxes allow multiline. Unfortunately textboxes do not allow me to add a checkbox or radiobutton beside each entry, so I added a combobox beneath the textbox. this allows the user to choose the entry they want to delete based on the date it was posted. when the user submits  a journal entry, the textbox updates to show the most recent previous entries first.